Extract out default and inline worker executors
